toolkit/content/tests/chrome/test_menulist_position.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="Menulist position Test"
        onload="setTimeout(init, 0)"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

<!--
  This test checks the position of a menulist's popup.
  -->

<script>
<![CDATA[
SimpleTest.waitForExplicitFinish();

var menulist;

function init()
{
  menulist = document.getElementById("menulist");
  menulist.open = true;
}

function isWithinHalfPixel(a, b)
{
  return Math.abs(a - b) <= 0.5;
}

const ismac = navigator.platform.indexOf("Mac") == 0;
function inputMargin(el) {
  let cs = getComputedStyle(el);
  // XXX Internal properties are not exposed in getComputedStyle, so we have to
  // use margin and rely on our knowledge of them matching negative margins
  // where appropriate.
  // return parseFloat(cs.getPropertyValue("-moz-window-input-region-margin"));
  return ismac ? 0 : Math.max(-parseFloat(cs.marginLeft), 0);
}

function popupShown()
{
  var menurect = menulist.getBoundingClientRect();
  var popuprect = menulist.menupopup.getBoundingClientRect();

  let marginLeft = parseFloat(getComputedStyle(menulist.menupopup).marginLeft);
  ok(isWithinHalfPixel(menurect.left + marginLeft, popuprect.left), `left position: ${menurect.left}, ${popuprect.left}`);
  ok(isWithinHalfPixel(menurect.right + marginLeft + 2 * inputMargin(menulist.menupopup), popuprect.right), `right position: ${menurect.right}, ${popuprect.right}`);

  let index = menulist.selectedIndex;
  if (menulist.selectedItem && navigator.platform.includes("Mac")) {
    let menulistlabel = menulist.shadowRoot.getElementById("label");
    let mitemlabel = menulist.selectedItem.querySelector(".menu-iconic-text");
    let paddingTop = parseFloat(getComputedStyle(menulist.selectedItem).paddingTop);

    ok(isWithinHalfPixel(menulistlabel.getBoundingClientRect().left,
                         mitemlabel.getBoundingClientRect().left),
       "Labels horizontally aligned for index " + index);
    ok(isWithinHalfPixel(menulistlabel.getBoundingClientRect().top + paddingTop,
                         mitemlabel.getBoundingClientRect().top),
       "Labels vertically aligned for index " + index);

    // Store the current value and reset it afterwards.
    let current = menulist.selectedIndex;

    // Cycle through the items to ensure that the popup doesn't move when the selection changes.
    for (let i = 0; i < menulist.itemCount; i++) {
      menulist.selectedIndex = i;

      let newpopuprect = menulist.menupopup.getBoundingClientRect();
      is(newpopuprect.x, popuprect.x, "Popup remained horizontally for index " + i + " starting at " + current);
      is(newpopuprect.y, popuprect.y, "Popup remained vertically for index " + i + " starting at " + current);
    }
    menulist.selectedIndex = current;
  }
  else {
    let marginTop = parseFloat(getComputedStyle(menulist.menupopup).marginTop);
    ok(isWithinHalfPixel(menurect.bottom + marginTop, popuprect.top),
       "Vertical alignment with no selection for index " + index);
  }

  menulist.open = false;
}

function popupHidden()
{
  if (!menulist.selectedItem) {
    SimpleTest.finish();
  }
  else {
    menulist.selectedItem = menulist.selectedItem.nextSibling;
    menulist.open = true;
  }
}
]]>
</script>

<hbox align="center" pack="center" style="margin-top: 140px;">
  <menulist id="menulist" onpopupshown="popupShown();" onpopuphidden="popupHidden();" native="true">
    <menupopup style="max-height: 90px;">
      <menuitem label="One"/>
      <menuitem label="Two"/>
      <menuitem label="Three"/>
      <menuitem label="Four"/>
      <menuitem label="Five"/>
      <menuitem label="Six"/>
      <menuitem label="Seven"/>
    </menupopup>
  </menulist>
</hbox>

<body xmlns="http://www.w3.org/1999/xhtml">
<p id="display">
</p>
<div id="content" style="display: none">
</div>
<pre id="test">
</pre>
</body>

</window>
